fixed some isssues for creating registration form

+++ src/main/java/com/server/HealthNet/Service/DoctorService.java
@@ -16,7 +16,7 @@ public class DoctorService {
         this.doctorRepository = doctorRepository;
     }
 
-    public int saveDoctor(Doctor doctor) {
+    public Long saveDoctor(Doctor doctor) {
         return doctorRepository.saveDoctor(doctor);
     }
