+++ src/main/java/com/server/HealthNet/Service/StaffService.java
@@ -30,7 +30,7 @@ public class StaffService {
         return staffRepository.updateStaff(staff);
     }
 
-    public int addStaff(Staff staff) {
+    public Long addStaff(Staff staff) {
         return staffRepository.saveStaff(staff);
     }
 }
